some debugs

+++ ft_strncpy.c
@@ -12,9 +12,9 @@
 
 #include "libft.h"
 
-char	*ft_strncpy(char *dst, const char *src, size_t len)
+char		*ft_strncpy(char *dst, const char *src, size_t len)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (src[i] && i < len)
